timestamps

+++ src/routers/task.js
@@ -5,9 +5,11 @@ const router = new express.Router();
 const app = express()
 app.use(router)
 
-router.get('/tasks', async (req, res) => {
+
+
+router.get('/tasks', auth, async (req, res) => {
     try{
-        const tasks = await Task.find({})
+        const tasks = await Task.find({owner: req.user._id })
         if(!tasks){
             return res.status(500).send()
         }
@@ -22,14 +24,14 @@ router.get('/tasks', async (req, res) => {
     }) */
 })
 
-router.get('/tasks/:id', async (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id
     try{
-        
-        const task = await Task.findOne({_id})
+        const task = await Task.findOne({ _id, owner: req.user._id })
+       
         if(!task){
             console.log("no task found")
-            return res.status(500).send()
+            return res.status(404).send()
         }
         res.send(task)
     } catch(e){
@@ -44,7 +46,7 @@ router.get('/tasks/:id', async (req, res) => {
 
 
 
-router.patch('/tasks/:id', async (req, res) => {
+router.patch('/tasks/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body);
     const allowedUpdates = ['description', 'completed'];
     const isValidOperation = updates.every((update) => allowedUpdates.includes(update));
@@ -54,13 +56,16 @@ router.patch('/tasks/:id', async (req, res) => {
     }
     
     try{
-        const task = await Task.findById(req.params.id)
-        updates.forEach((update) => task[update] = req.body[update])
-        await task.save()
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id })
+       
+       
         //const task = await Task.findByIdAndUpdate(_id, req.body, { new: true, runValidators: true})
         if(!task){
             res.status(404).send();
         }
+
+        updates.forEach((update) => task[update] = req.body[update])
+        await task.save()
         res.send(task);
     } catch(e){
         console.log(e);
@@ -92,9 +97,9 @@ router.post('/tasks', auth, async (req, res) =>{
 
 
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id',auth, async (req, res) => {
     try{
-        const task = await Task.findByIdAndDelete(req.params.id);
+        const task = await Task.findByIdAndDelete({ _id: req.params.id, owner: req.user._id });
         if(!task){
             res.status(404).send()
         }
